backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database  (psycopg2-binary not needed — asyncpg covers everything)
sqlalchemy[asyncio]==2.0.36        # 2.0.36+ required for Python 3.13
asyncpg==0.31.0        # 0.31+ required for Python 3.13
alembic==1.13.1

# Settings & validation
pydantic==2.11.10      # 2.7.x doesn't build pydantic-core on Py 3.13
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth & crypto
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1          # pin to last version compatible with passlib
ecdsa==0.19.2
pyshamir==1.0.4        # API: split(bytes,n,k) -> list[bytearray], combine(list[bytearray]) -> bytes

# Blockchain — Base Sepolia testnet (chain ID 84532)
web3==7.6.0            # 6.x fails to build ckzg on Windows without a C compiler

# File upload handling
python-multipart==0.0.9

# QR code & PDF (Module 12)
qrcode[pil]==7.4.2
reportlab==4.1.0
Pillow>=11.0.0         # 10.x doesn't build on Python 3.13

# v4: ECIES VC encryption
cryptography>=42.0.0

# v4: IPFS / Pinata HTTP client
httpx>=0.27.0

# Env
python-dotenv==1.0.1
